Migrated to a cpp file
