+++ Tests/CMakeLists.txt
@@ -3,7 +3,7 @@ add_subdirectory(1DCondition)
 #add_subdirectory(Recurse)
 add_subdirectory(FunctionCall)
 add_subdirectory(bubbleSort)
-add_subdirectory(Threads)
+#add_subdirectory(Threads)
 if (${ENABLE_TESTING_LONG})
     add_subdirectory(2DConv)
     add_subdirectory(MatrixMultiply)
